fix: перенести Notyfyer в notifyloop
