+++ internal/notifyloop/messages-spawn.go
@@ -1,4 +1,4 @@
-package app
+package notifyloop
 
 import (
 	"context"
@@ -6,39 +6,38 @@ import (
 	"gitlab.com/thefrol/notty/internal/entity"
 )
 
-// Notifyerrrr представляет слой приложения
-// для отправщика. Тут функции генерации сообщения
-// и отправки сообщений
-type Notifyerrrr struct {
+// Notifyer это уточнение над репозиторием сообщений, который умеет хорошо
+// генерировать сообщения
+type Notifyer struct {
 	messages Messager
 	sender   Sender
 }
 
-func NewNotifyerrrr(messages Messager, sender Sender) *Notifyerrrr {
-	return &Notifyerrrr{
+func NewNotifyer(messages Messager, sender Sender) *Notifyer {
+	return &Notifyer{
 		messages: messages,
 		sender:   sender,
 	}
 }
 
 // CreateMessages создает n сообщений, или меньше.
-func (notty Notifyerrrr) CreateMessages(ctx context.Context, n int) ([]entity.Message, error) {
+func (notty Notifyer) CreateMessages(ctx context.Context, n int) ([]entity.Message, error) {
 	return notty.messages.LockedSpawn(ctx, n, entity.StatusInitial)
 }
 
 // ReserveMessages резервирует n сообщений с изначальным статусом
 // fromStatus
-func (notty Notifyerrrr) ReserveMessages(ctx context.Context, n int, fromStatus string) ([]entity.Message, error) {
+func (notty Notifyer) ReserveMessages(ctx context.Context, n int, fromStatus string) ([]entity.Message, error) {
 	return notty.messages.ReserveFromStatus(ctx, n, fromStatus)
 }
 
 // SendSMS отправляет смской сообщение m
-func (notty Notifyerrrr) SendSMS(ctx context.Context, m entity.Message) error {
+func (notty Notifyer) SendSMS(ctx context.Context, m entity.Message) error {
 	return notty.sender.Send(ctx, m)
 }
 
 // UpdateMessage обновляеят статус у сообщения
-func (notty Notifyerrrr) UpdateMessage(ctx context.Context, m entity.Message) error {
+func (notty Notifyer) UpdateMessage(ctx context.Context, m entity.Message) error {
 	_, err := notty.messages.Update(ctx, m)
 	if err != nil {
 		return err
